+++ shared/src/commonMain/kotlin/ui/views/CredentialDetailsView.kt
@@ -1,7 +1,5 @@
 package ui.views
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -12,7 +10,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CheckCircle
@@ -20,18 +17,14 @@ import androidx.compose.material.icons.filled.Warning
 import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.unit.dp
 import at.asitplus.valera.resources.Res
@@ -49,11 +42,14 @@ import ui.composables.Logo
 import ui.composables.PersonAttributeDetailCardHeading
 import ui.composables.PersonAttributeDetailCardHeadingIcon
 import ui.composables.ScreenHeading
+import ui.composables.TrustState
 import ui.composables.buttons.NavigateUpButton
 import ui.composables.credentials.EuPidCredentialView
 import ui.composables.credentials.GenericCredentialSummaryCardContent
 import ui.composables.credentials.GenericMetadataCredentialView
 import ui.composables.credentials.MobileDrivingLicenceCredentialView
+import ui.composables.TrustStatusBanner
+import ui.composables.credentials.*
 import ui.models.CredentialFreshnessSummaryUiModel
 import ui.models.ResolvedCredential
 import ui.viewmodels.CredentialDetailsViewModel
@@ -177,54 +173,5 @@ fun CredentialDetailsSummaryView(
         // No extra horizontal padding here: the enclosing Column already applies it, otherwise the cards
         // (technical data, status, contents, cnf) would be indented twice.
         GenericCredentialSummaryCardContent(credential = credential)
-    }
-}
-
-
-
-enum class TrustState {
-    TRUSTED, UNTRUSTED, UNKNOWN, EVALUATING
-}
-
-@Composable
-fun TrustStatusBanner(trustState: TrustState, modifier: Modifier = Modifier) {
-    val (backgroundColor, contentColor, icon, text) = when (trustState) {
-        TrustState.TRUSTED -> listOf(
-            Color(0xFFE8F5E9), Color(0xFF2E7D32),
-            Icons.Filled.CheckCircle, "Trusted Issuer"
-        )
-        TrustState.UNTRUSTED -> listOf(
-            Color(0xFFFFEBEE), Color(0xFFC62828),
-            Icons.Filled.Error, "Untrusted Issuer"
-        )
-        TrustState.UNKNOWN -> listOf(
-            Color(0xFFFFF8E1), Color(0xFFF57F17),
-            Icons.Filled.Warning, "Trust Status Unknown"
-        )
-        TrustState.EVALUATING -> listOf(
-            MaterialTheme.colorScheme.surfaceVariant, MaterialTheme.colorScheme.onSurfaceVariant,
-            Icons.Filled.Warning, "Evaluating Trust..."
-        )
-    }
-
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(8.dp))
-            .background(backgroundColor as Color)
-            .padding(12.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            imageVector = icon as androidx.compose.ui.graphics.vector.ImageVector,
-            contentDescription = null,
-            tint = contentColor as Color
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        Text(
-            text = text as String,
-            color = contentColor,
-            style = MaterialTheme.typography.bodyMedium
-        )
     }
 }
